+++ src/main/java/wooteco/subway/domain/line/Line.java
@@ -1,8 +1,13 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.line;
 
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
+
+import wooteco.subway.domain.Name;
+import wooteco.subway.domain.line.section.Section;
+import wooteco.subway.domain.line.section.Sections;
+import wooteco.subway.domain.Station;
 
 public class Line {
