+++ src/main/java/wooteco/subway/domain/path/SubwayShortestPath.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.path;
 
 import java.util.HashMap;
 import java.util.List;
@@ -10,6 +10,10 @@ import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
+
+import wooteco.subway.domain.line.Line;
+import wooteco.subway.domain.line.section.Section;
+import wooteco.subway.domain.Station;
 
 public class SubwayShortestPath {
