+++ src/main/java/wooteco/subway/repository/JdbcLineRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 import wooteco.subway.repository.dao.LineDao;
 import wooteco.subway.repository.table.LineTable;
-import wooteco.subway.domain.Line;
+import wooteco.subway.domain.line.Line;
 
 @Repository
 public class JdbcLineRepository implements LineRepository {
@@ -29,11 +29,10 @@ public class JdbcLineRepository implements LineRepository {
 
 	@Override
 	public List<Line> findAll() {
-		List<Line> collect = lineDao.findAll().stream()
+		return lineDao.findAll().stream()
 			.map(LineTable::toEntity)
 			.map(line -> line.createWithSection(sectionRepository.findByLineId(line.getId())))
 			.collect(Collectors.toList());
-		return collect;
 	}
 
 	@Override
