+++ src/main/java/wooteco/subway/repository/SectionRepository.java
@@ -2,7 +2,7 @@ package wooteco.subway.repository;
 
 import java.util.List;
 
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.line.section.Section;
 
 public interface SectionRepository {
 	Long save(Long lineId, Section section);
