+++ src/main/java/wooteco/subway/service/LineService.java
@@ -7,8 +7,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import wooteco.subway.repository.LineRepository;
 import wooteco.subway.repository.SectionRepository;
-import wooteco.subway.domain.Line;
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.line.Line;
+import wooteco.subway.domain.line.section.Section;
 import wooteco.subway.repository.SectionsDirtyChecker;
 
 @Service
